[42] Prevent calling 'onSuccess' callback in @socketio.on_error
When error was not reraised in on_error handler then
socketio.emit callback was fired (with null arguments).
When error was reraised then whole socket was closed.
It was necessary to add special arg sent to client which
specifies whether callback should be called or not.
It was done by monkey-patching socketio.emit
Started to use varargs for emit function arguments.