[116] Add a death action for animals, fix setup of sample dom. animals
A new action for animals (locations and items) which is run in case of 1.0 damage. Added dead_horse, dead_aurochs and dead_pig types
A fix of setup used by a development server is turning a horse into a mare (female horse), adding correct dead types and adding missing ANIMAL/DOMESTICATED properties for a pig and a mare.